requirements.txt
pandas>=2.0
pydantic>=2.0
google-genai>=1.0
langdetect>=1.0.9
python-dotenv>=1.0
PyYAML>=6.0
